Clean up management commands with Client and ArticlePipeline
